[release-only] Revert "[routing] Pedestrian directions engine."
This reverts commit 8e0f81f4f9896c4434a9e5a673fd67ee424b8f67.
