+++ routing/car_directions.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "routing/directions_engine.hpp"
-#include "routing/directions_engine_helpers.hpp"
 #include "routing/index_road_graph.hpp"
 #include "routing/loaded_path_segment.hpp"
 #include "routing/segment.hpp"
@@ -22,6 +21,17 @@ namespace routing
 class CarDirectionsEngine : public IDirectionsEngine
 {
 public:
+  struct AdjacentEdges
+  {
+    explicit AdjacentEdges(size_t ingoingTurnsCount = 0) : m_ingoingTurnsCount(ingoingTurnsCount) {}
+    bool IsAlmostEqual(AdjacentEdges const & rhs) const;
+
+    turns::TurnCandidates m_outgoingTurns;
+    size_t m_ingoingTurnsCount;
+  };
+
+  using AdjacentEdgesMap = std::map<SegmentRange, AdjacentEdges>;
+
   CarDirectionsEngine(DataSource const & dataSource, std::shared_ptr<NumMwmIds> numMwmIds);
 
   // IDirectionsEngine override:
@@ -30,5 +40,30 @@ public:
                 Route::TStreets & streetNames,
                 std::vector<geometry::PointWithAltitude> & routeGeometry,
                 std::vector<Segment> & segments) override;
+  void Clear() override;
+
+private:
+  FeaturesLoaderGuard & GetLoader(MwmSet::MwmId const & id);
+  void LoadPathAttributes(FeatureID const & featureId, LoadedPathSegment & pathSegment);
+  void GetSegmentRangeAndAdjacentEdges(IRoadGraph::EdgeVector const & outgoingEdges,
+                                       Edge const & inEdge, uint32_t startSegId, uint32_t endSegId,
+                                       SegmentRange & segmentRange,
+                                       turns::TurnCandidates & outgoingTurns);
+  /// \brief The method gathers sequence of segments according to IsJoint() method
+  /// and fills |m_adjacentEdges| and |m_pathSegments|.
+  void FillPathSegmentsAndAdjacentEdgesMap(IndexRoadGraph const & graph,
+                                           std::vector<geometry::PointWithAltitude> const & path,
+                                           IRoadGraph::EdgeVector const & routeEdges,
+                                           base::Cancellable const & cancellable);
+
+  void GetEdges(IndexRoadGraph const & graph, geometry::PointWithAltitude const & currJunction,
+                bool isCurrJunctionFinish, IRoadGraph::EdgeVector & outgoing,
+                IRoadGraph::EdgeVector & ingoing);
+
+  AdjacentEdgesMap m_adjacentEdges;
+  TUnpackedPathSegments m_pathSegments;
+  DataSource const & m_dataSource;
+  std::shared_ptr<NumMwmIds> m_numMwmIds;
+  std::unique_ptr<FeaturesLoaderGuard> m_loader;
 };
 }  // namespace routing
