+++ routing/directions_engine.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "routing/directions_engine_helpers.hpp"
 #include "routing/index_road_graph.hpp"
 #include "routing/road_point.hpp"
 #include "routing/route.hpp"
@@ -8,14 +7,10 @@
 
 #include "traffic/traffic_info.hpp"
 
-#include "indexer/data_source.hpp"
-
 #include "geometry/point_with_altitude.hpp"
 
-#include "base/assert.hpp"
 #include "base/cancellable.hpp"
 
-#include <memory>
 #include <vector>
 
 namespace routing
@@ -23,12 +18,6 @@ namespace routing
 class IDirectionsEngine
 {
 public:
-  IDirectionsEngine(DataSource const & dataSource, std::shared_ptr<NumMwmIds> numMwmIds)
-    : m_dataSource(dataSource), m_numMwmIds(numMwmIds)
-  {
-    CHECK(m_numMwmIds, ());
-  }
-
   virtual ~IDirectionsEngine() = default;
 
   // @TODO(bykoianko) Method Generate() should fill
@@ -42,31 +31,6 @@ public:
                         Route::TStreets & streetNames,
                         std::vector<geometry::PointWithAltitude> & routeGeometry,
                         std::vector<Segment> & segments) = 0;
-  void Clear();
-
-protected:
-  FeaturesLoaderGuard & GetLoader(MwmSet::MwmId const & id);
-  void LoadPathAttributes(FeatureID const & featureId, LoadedPathSegment & pathSegment);
-  void GetSegmentRangeAndAdjacentEdges(IRoadGraph::EdgeVector const & outgoingEdges,
-                                       Edge const & inEdge, uint32_t startSegId, uint32_t endSegId,
-                                       SegmentRange & segmentRange,
-                                       turns::TurnCandidates & outgoingTurns);
-  /// \brief The method gathers sequence of segments according to IsJoint() method
-  /// and fills |m_adjacentEdges| and |m_pathSegments|.
-  void FillPathSegmentsAndAdjacentEdgesMap(IndexRoadGraph const & graph,
-                                           std::vector<geometry::PointWithAltitude> const & path,
-                                           IRoadGraph::EdgeVector const & routeEdges,
-                                           base::Cancellable const & cancellable);
-
-  void GetEdges(IndexRoadGraph const & graph, geometry::PointWithAltitude const & currJunction,
-                bool isCurrJunctionFinish, IRoadGraph::EdgeVector & outgoing,
-                IRoadGraph::EdgeVector & ingoing);
-
-  AdjacentEdgesMap m_adjacentEdges;
-  TUnpackedPathSegments m_pathSegments;
-
-  DataSource const & m_dataSource;
-  std::shared_ptr<NumMwmIds> m_numMwmIds;
-  std::unique_ptr<FeaturesLoaderGuard> m_loader;
+  virtual void Clear() = 0;
 };
 }  // namespace routing
