+++ routing/pedestrian_directions.cpp
@@ -1,7 +1,11 @@
 #include "routing/pedestrian_directions.hpp"
 
 #include "routing/road_graph.hpp"
-#include "routing/turns_generator.hpp"
+#include "routing/routing_helpers.hpp"
+
+#include "indexer/classificator.hpp"
+#include "indexer/feature.hpp"
+#include "indexer/ftypes_matcher.hpp"
 
 #include "base/assert.hpp"
 #include "base/logging.hpp"
@@ -10,11 +14,28 @@
 
 using namespace std;
 
+namespace
+{
+bool HasType(uint32_t type, feature::TypesHolder const & types)
+{
+  for (uint32_t t : types)
+  {
+    t = ftypes::BaseChecker::PrepareToMatch(t, 2);
+    if (type == t)
+      return true;
+  }
+  return false;
+}
+}  // namespace
+
 namespace routing
 {
-PedestrianDirectionsEngine::PedestrianDirectionsEngine(DataSource const & dataSource,
-                                                       shared_ptr<NumMwmIds> numMwmIds)
-  : IDirectionsEngine(dataSource, move(numMwmIds))
+
+PedestrianDirectionsEngine::PedestrianDirectionsEngine(shared_ptr<NumMwmIds> numMwmIds)
+  : m_typeSteps(classif().GetTypeByPath({"highway", "steps"}))
+  , m_typeLiftGate(classif().GetTypeByPath({"barrier", "lift_gate"}))
+  , m_typeGate(classif().GetTypeByPath({"barrier", "gate"}))
+  , m_numMwmIds(move(numMwmIds))
 {
 }
 
@@ -25,47 +46,60 @@ bool PedestrianDirectionsEngine::Generate(IndexRoadGraph const & graph,
                                           vector<geometry::PointWithAltitude> & routeGeometry,
                                           vector<Segment> & segments)
 {
-  CHECK(m_numMwmIds, ());
-
-  m_adjacentEdges.clear();
-  m_pathSegments.clear();
   turns.clear();
   streetNames.clear();
   segments.clear();
-  routeGeometry.clear();
+  routeGeometry = path;
 
+  // Note. According to Route::IsValid() method route of zero or one point is invalid.
   if (path.size() <= 1)
     return false;
-
-  size_t const pathSize = path.size();
 
   vector<Edge> routeEdges;
   graph.GetRouteEdges(routeEdges);
 
-  if (routeEdges.empty())
-    return false;
+  CalculateTurns(graph, routeEdges, turns, cancellable);
 
-  if (cancellable.IsCancelled())
-    return false;
-
-  FillPathSegmentsAndAdjacentEdgesMap(graph, path, routeEdges, cancellable);
-
-  if (cancellable.IsCancelled())
-    return false;
-
-  RoutingEngineResult resultGraph(routeEdges, m_adjacentEdges, m_pathSegments);
-  auto const res = MakeTurnAnnotationPedestrian(resultGraph, *m_numMwmIds, cancellable,
-                                                routeGeometry, turns, streetNames, segments);
-
-  if (res != RouterResultCode::NoError)
-    return false;
-
-  CHECK_EQUAL(
-      routeGeometry.size(), pathSize,
-      ("routeGeometry and path have different sizes. routeGeometry size:", routeGeometry.size(),
-       "path size:", pathSize, "segments size:", segments.size(), "routeEdges size:",
-       routeEdges.size(), "resultGraph.GetSegments() size:", resultGraph.GetSegments().size()));
-
+  graph.GetRouteSegments(segments);
   return true;
+}
+
+void PedestrianDirectionsEngine::CalculateTurns(IndexRoadGraph const & graph,
+                                                vector<Edge> const & routeEdges,
+                                                Route::TTurns & turns,
+                                                base::Cancellable const & cancellable) const
+{
+  for (size_t i = 0; i < routeEdges.size(); ++i)
+  {
+    if (cancellable.IsCancelled())
+      return;
+
+    Edge const & edge = routeEdges[i];
+
+    feature::TypesHolder types;
+    graph.GetEdgeTypes(edge, types);
+
+    if (HasType(m_typeSteps, types))
+    {
+      if (edge.IsForward())
+        turns.emplace_back(i, turns::PedestrianDirection::Upstairs);
+      else
+        turns.emplace_back(i, turns::PedestrianDirection::Downstairs);
+    }
+    else
+    {
+      graph.GetJunctionTypes(edge.GetStartJunction(), types);
+
+      // @TODO(bykoianko) Turn types Gate and LiftGate should be removed.
+      if (HasType(m_typeLiftGate, types))
+        turns.emplace_back(i, turns::PedestrianDirection::LiftGate);
+      else if (HasType(m_typeGate, types))
+        turns.emplace_back(i, turns::PedestrianDirection::Gate);
+    }
+  }
+
+  // direction "arrival"
+  // (index of last junction is the same as number of edges)
+  turns.emplace_back(routeEdges.size(), turns::PedestrianDirection::ReachedYourDestination);
 }
 }  // namespace routing
