+++ routing/pedestrian_directions.hpp
@@ -1,10 +1,6 @@
 #pragma once
 
 #include "routing/directions_engine.hpp"
-#include "routing/directions_engine_helpers.hpp"
-#include "routing/loaded_path_segment.hpp"
-#include "routing/routing_result_graph.hpp"
-#include "routing/turn_candidate.hpp"
 
 #include "routing_common/num_mwm_id.hpp"
 
@@ -15,10 +11,11 @@
 
 namespace routing
 {
+
 class PedestrianDirectionsEngine : public IDirectionsEngine
 {
 public:
-  PedestrianDirectionsEngine(DataSource const & dataSource, std::shared_ptr<NumMwmIds> numMwmIds);
+  PedestrianDirectionsEngine(std::shared_ptr<NumMwmIds> numMwmIds);
 
   // IDirectionsEngine override:
   bool Generate(IndexRoadGraph const & graph, std::vector<geometry::PointWithAltitude> const & path,
@@ -26,5 +23,16 @@ public:
                 Route::TStreets & streetNames,
                 std::vector<geometry::PointWithAltitude> & routeGeometry,
                 std::vector<Segment> & segments) override;
+  void Clear() override {}
+
+private:
+  void CalculateTurns(IndexRoadGraph const & graph, std::vector<Edge> const & routeEdges,
+                      Route::TTurns & turnsDir, base::Cancellable const & cancellable) const;
+
+  uint32_t const m_typeSteps;
+  uint32_t const m_typeLiftGate;
+  uint32_t const m_typeGate;
+  std::shared_ptr<NumMwmIds> const m_numMwmIds;
 };
+
 }  // namespace routing
