+++ routing/route.cpp
@@ -203,8 +203,7 @@ bool Route::GetNextTurn(double & distanceToTurnMeters, TurnItem & nextTurn) cons
   // |curIdx| + 1 - 1 is an index of segment to start look for the closest turn.
   GetClosestTurn(curIdx, curTurn);
   CHECK_LESS(curIdx, curTurn.m_index, ());
-  if (curTurn.m_turn == CarDirection::ReachedYourDestination ||
-      curTurn.m_pedestrianTurn == PedestrianDirection::ReachedYourDestination)
+  if (curTurn.m_turn == CarDirection::ReachedYourDestination)
   {
     nextTurn = TurnItem();
     return false;
